LOW: attempt to fix issue with cast to Double in Swing/Linux

+++ gina-swing/src/main/java/org/openflexo/gina/swing/view/widget/JFIBNumberWidget.java
@@ -257,7 +257,7 @@ public class JFIBNumberWidget<T extends Number> extends FIBNumberWidgetImpl<Numb
 		public abstract T getEditedValue();
 
 		public void setEditedValue(T aValue) {
-			valueChooser.setValue(aValue);
+			valueChooser.setValue(aValue.doubleValue());
 		}
 
 		public JSpinner getValueChooser() {
